thermal: rockchip: Disable auto mode before reset and initialize
There is a issuse when change auto mode to user mode, and now the
SRST_TSADC only reset the tsadc clk, don't reset regsters, so when
initialize tsadc the mode will be changed from auto to user, the tsadc
will work abnormally. Change mode to user and then do SRST_TSADC is
a workaround, and add SRST_TSADC_P and SRST_TSADC resets is a better way,
we will add another patch.

Change-Id: Iaed86545a5cb0c85250641b27f1aae0a1bcf9797
Signed-off-by: Finley Xiao <finley.xiao@rock-chips.com>
Reviewed-on: https://10.10.10.29/c/rk/kernel-stable/+/135492
Reviewed-by: Jianhong Chen <chenjh@rock-chips.com>
Reviewed-by: Lin JIan Hua <linjh@rock-chips.com>
Reviewed-by: Hans Yang <yhx@rock-chips.com>
Tested-by: Hans Yang <yhx@rock-chips.com>