rk3399pro-firefly-aioc: NPU: Optimize the transmission speed of RK1808 using PCIE bus
Reduce the DMA BUF of PCIE from the original 1M to 128K
to optimize the transmission speed of RK1808 using PCIE bus.
As a result, the inference speed of the NPU RK1808 integrated
in the RK339PRO is increased by 60%, and the continuous reading
and writing speed of small files using the PCIE interface will increase,
but the reading and writing speed of large files will decrease.

If you want to use this feature, you need to modify dts and config.
And the version of firefly-3399pronpu-driver must be at least 1.7.1 installed on Ubuntu or Debian running on the board.

E.g:
--- a/arch/arm64/boot/dts/rockchip/rk3399pro-firefly-aioc.dts
+++ b/arch/arm64/boot/dts/rockchip/rk3399pro-firefly-aioc.dts

+        &PCIE_RK1808_OPTIMIZATION {
+               status = "okay";
+       };

--- a/arch/arm64/configs/firefly3399pro_linux_defconfig
+++ b/arch/arm64/configs/firefly3399pro_linux_defconfig
@@ -40,7 +40,7 @@ CONFIG_PCI=y
 CONFIG_PCI_MSI=y
 CONFIG_ROCKCHIP_PCIE_DMA_OBJ=y
 CONFIG_PCIE_ROCKCHIP=y
-CONFIG_PCIE_RK1808_OPTIMIZATION=n
+CONFIG_PCIE_RK1808_OPTIMIZATION=y

Ubuntu or Debian:
apt update
apt install firefly-3399pronpu-driver